refa before input val

+++ my-swr-app/src/presentation/message/useMessageActions.ts
@@ -1,15 +1,15 @@
 import { useCallback } from 'react';
-import { MessagePresenter } from '..'; // Или правильный путь
-import type { Message } from '../../Domain/Entities/MessageTypes'; // Импорт типа Message
+import { MessagePresenter } from '..'; 
+import type { Message } from '../../Domain/Entities/MessageTypes'; 
 
 export const useMessageActions = (
   setMessages: React.Dispatch<React.SetStateAction<Message[]>>
 ) => {
   const deleteMessage = useCallback(async (messageId: string) => {
-    // if (messageId.startsWith('temp-')) {
-    //   alert('Нельзя удалить неотправленное сообщение');
-    //   return;
-    // }
+    if (messageId.startsWith('temp-')) {
+      alert('Нельзя удалить неотправленное сообщение');
+      return;
+    }
     try {
       await MessagePresenter.deleteMessageById(messageId);
       setMessages(prev => prev.filter(m => m.id !== messageId));
@@ -17,7 +17,7 @@ export const useMessageActions = (
       console.log('Failed to delete message', error);
       alert('Failed to delete message');
     }
-  }, [setMessages]); // Добавь setMessages в зависимости
+  }, [setMessages]); 
 
   return { deleteMessage };
 };
